John Fay & James Jones - Added puScrollBar, major changes to puLargeInput, passive motion highlighting (requires puMouse to be called for passive movements)

+++ trunk/src/pui/puTriSlider.cxx
@@ -43,12 +43,12 @@ void puTriSlider::draw ( int dx, int dy )
 
     float val;
     if ( getMaxValue() > getMinValue() )
-      val = (float)(getIntegerValue () - getMinValue()) / (float)(getMaxValue() - getMinValue()) ;
+      val = (float)(getFloatValue () - getMinValue()) / (float)(getMaxValue() - getMinValue()) ;
     else
       val = 0.5f ;
 
     char str_value[10] ;
-    sprintf (str_value, "%d", getIntegerValue () ) ;
+    sprintf (str_value, "%d", getFloatValue () ) ;
 
     draw_slider_box ( dx, dy, abox, val, str_value ) ;
 
@@ -62,7 +62,7 @@ void puTriSlider::draw ( int dx, int dy )
     else
       val = 1.0f ;
 
-    sprintf (str_value, "%d", getCurrentMax() ) ;
+    sprintf (str_value, "%g", getCurrentMax() ) ;
 
     draw_slider_box ( dx, dy, abox, val, str_value ) ;
 
@@ -73,7 +73,7 @@ void puTriSlider::draw ( int dx, int dy )
     else
       val = 0.0f ;
 
-    sprintf (str_value, "%d", getCurrentMin() ) ;
+    sprintf (str_value, "%g", getCurrentMin() ) ;
 
     draw_slider_box ( dx, dy, abox, val, str_value ) ;
 
@@ -139,12 +139,12 @@ void puTriSlider::doHit ( int button, int updown, int x, int y )
     {
       if ( getActiveButton() == 0 )  // No currently-active slider, set whichever is closest
       {
-        if ( (new_value-getCurrentMin()) < (getIntegerValue()-new_value) ) // Closest to current_min
+        if ( (new_value-getCurrentMin()) < (getFloatValue()-new_value) ) // Closest to current_min
         {
           setCurrentMin ( checkStep(new_value) ) ;
           setActiveButton ( 1 ) ;
         }
-        else if ( (new_value-getIntegerValue()) > (getCurrentMax()-new_value) ) // Closest to current_max
+        else if ( (new_value-getFloatValue()) > (getCurrentMax()-new_value) ) // Closest to current_max
         {
           setCurrentMax ( checkStep(new_value) ) ;
           setActiveButton ( 3 ) ;
@@ -158,7 +158,7 @@ void puTriSlider::doHit ( int button, int updown, int x, int y )
       else if ( getActiveButton() == 1 )  // Currently moving current_min
       {
         setCurrentMin ( checkStep(new_value) ) ;
-        if ( getIntegerValue() < getCurrentMin() ) setValue ( getCurrentMin() ) ;
+        if ( getFloatValue() < getCurrentMin() ) setValue ( getCurrentMin() ) ;
         if ( getCurrentMax() < getCurrentMin() ) setCurrentMax ( getCurrentMin() ) ;
       }
       else if ( getActiveButton() == 2 )  // Currently moving central value
@@ -170,7 +170,7 @@ void puTriSlider::doHit ( int button, int updown, int x, int y )
       else if ( getActiveButton() == 3 )  // Currently moving current_max
       {
         setCurrentMax ( checkStep(new_value) ) ;
-        if ( getIntegerValue() > getCurrentMax() ) setValue ( getCurrentMax() ) ;
+        if ( getFloatValue() > getCurrentMax() ) setValue ( getCurrentMax() ) ;
         if ( getCurrentMax() < getCurrentMin() ) setCurrentMin ( getCurrentMax() ) ;
       }
     }
